+++ CHANGELOG.md
@@ -4,5 +4,8 @@ HTTP-Header
 Changelog
 ---------
 
+<b>Version 1.1</b>
+- Anpassung für DarkMode
+
 <b>Version 1.0</b>
 - Erste Version mit allen grundlegenden Headern